Removed server out of params. Now reload just uses port. This allows the app to become simpler for many reasons:
* Allows for easy setup in an app
* Removes complex two part function for people who have split up app and server (this reverses #71)

This commit closes #95 as it allows the reload path to be configured, allowing someone to use the `reload` route in their application if they needed outside of reload.

This commit refactors the params again to start with the only required argument app, everything else has defaults and is in an optional object param.